Combining templates and views sections.
The docs seemed a bit weird insofar as there are two separate sections for templates/views, and they both say to do fairly different things. I've combined them here so that there's a very quick intro in "Templates" part of the "Getting to know Scalatra" section, and a longer discussion in the "Views" section.