Removing the Scalate stuff from "Helpers" section.
Scalate stuff is already in the "views" section, putting an additional bit down here in "Helpers" is confusing. 
I am not sure whether the code I've removed is showing anything super-important, it doesn't seem like it
to me. 

I am not sure if this example is supposed to be showing the use of a "notFound" method, I could move it up into
the "views" section maybe?